k8s configs and changes to the flow

+++ microservices/commons/src/main/java/com/safety/recognition/cassandra/model/CrimeCategory.java
@@ -19,7 +19,7 @@ public class CrimeCategory {
     private String url;
 
     @Column
-    private Long numericRepresentation;
+    private Integer numericRepresentation;
 
     @Column
     private String name;
